requirements.txt
nextcord
tortoise-orm
pydantic
typing_extensions
deepdiff
PyYAML
toml